Now using pointers to data instead of copy, also fixed some types from int to size_t

+++ Abel/head/Base/MVNumericalMethod.h
@@ -9,13 +9,13 @@ protected:
 	Eigen::VectorXd starting_point;
 	double tol;
 	double error;
-	int max_iter;
+	size_t max_iter;
 	int iter_counter;
 	Eigen::VectorXd result;
 	bool was_run;
 	bool hasStart;
 
-	MVNumericalMethod(size_t dimension, double tol = 0.01, int max_iter = 100, bool was_run = 0, int iter_counter = 0, double err = 0.0,
+	MVNumericalMethod(size_t dimension, double tol = 0.01, size_t max_iter = 100, bool was_run = 0, int iter_counter = 0, double err = 0.0,
 		const Eigen::VectorXd& res = Eigen::VectorXd(), const Eigen::VectorXd& starting_point = Eigen::VectorXd(), bool hasStart = false);
 	MVNumericalMethod& operator=(const MVNumericalMethod& nm);
