+++ Abel/head/OptimizationProblems/qprog_Result.h
@@ -6,8 +6,8 @@ struct qprog_Result {
 	double error;
 	size_t niter;
 	bool isDiverging;
-	int z_num; // number of dual variables for equality constraints
-	int l_num; // number of dual variables for inequalities
+	size_t z_num; // number of dual variables for equality constraints
+	size_t l_num; // number of dual variables for inequalities
 
-	qprog_Result(const Eigen::VectorXd& result, const Eigen::VectorXd& dual_variables, double error, size_t niter, bool isDiverging, int z_num, int l_num);
+	qprog_Result(const Eigen::VectorXd& result, const Eigen::VectorXd& dual_variables, double error, size_t niter, bool isDiverging, size_t z_num, size_t l_num);
 };
